Expose vagrant forwarded ports to public networks
Despite what the documents say here:
https://www.vagrantup.com/docs/networking/forwarded_ports.html

Vagrant seems to bind forwarded ports only to localhost when using libvirt.
Thats not ideal for use on remote systems, when you want to access the patchwork
and gitlab web interface from your local system, so augment the Vagrant file to
explicitly bind to INADDR_ANY, so that you can reach the vagrant boxes from
remote systems.  Its less secure, but for a development environment, that should
be fine.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>