Use the merge request iid instead of the global id
The merge request hook expects the merge id within the project, not the
Gitlab-global merge object id. This is stored in the "iid" field rather
than the "id" field of the web hook payload. It happened to work in
testing because there is only one project in the CI environment so id ==
iid.

Signed-off-by: Jeremy Cline <jcline@redhat.com>