Catch errors in record bridging to skip already sent email on retry
if we get an error halfway through sending a series via email, we need to retry,
but we don't want to resend already sent email.  Avoid this by using ValueError
to detect already bridged submissions and skip them.  Also, should we encounter
an error in email sending after recroding a bridged submission, delete that
submission from the db so that we properly retry it later

Signed-off-by: Neil Horman <nhorman@redhat.com>