Cc users to a merge request if tagged
This is a bit of a hack, but we need to be able to indicate that certain
users should be copied on the email version of the merge request. This
allows merge requests to be tagged with tags in the format "Cc: <email>"
to Cc that email on the merge request.

Signed-off-by: Jeremy Cline <jcline@redhat.com>