Move to Django fixtures for Series creation
The series ID is used in the branch ID which means it is in the HTTP
requests we record for tests. Rather than creating series per test which
may result in different series IDs depending on ordering, use the Django
test fixture loading feature to load a basic test environment with a
couple of series. More fixtures can be added with django-admin dumpdata.

Signed-off-by: Jeremy Cline <jcline@redhat.com>