Add a web hook handler for pipeline results
This handler is a bit hacky as there doesn't seem to be a great way to
map pipeline results to merge requests and documentation about the
attributes of the hook are sparse.

This filters out the pipeline hooks for those marked as "success" and
then searches open merge requests for pipelines that match the pipeline
ID in the hook. It ignores pull requests that have been marked as a work
in progress and merge requests that are currently unmerge-able.

Signed-off-by: Jeremy Cline <jcline@redhat.com>