Adjust the pipeline hook to use "merge_request_event" source
GitLab *does* include the merge request information if the pipeline is
triggered by a merge request event. One way to do this by tagging the
stage(s) with:

only:
  - merge_requests

This changes the hook to only deal with pipelines marked as being from
merge request events and greatly simplifies the mapping of a pipeline to
a merge request (because the hook payload includes it).

Signed-off-by: Jeremy Cline <jcline@redhat.com>