Drop the "From email" label check in the webhook view
The task that turns merge requests into emails checks the label since
it's also used for pipeline events and those events don't come with
label details. Although this will result in a couple more Celery tasks,
I believe it's worth it to have a single place where labels are checked.

Signed-off-by: Jeremy Cline <jcline@redhat.com>