Add "update" to merge request hook and filters to the task
Accept merge request updates in the web hook handler, which allows us to
send new series versions. We avoid sending duplicates by checking the
bridged submissions table for prior emails for the current merge request
head SHA. Furthermore, we add a way to wait on CI completing.

The reason for this rather than using the Pipelines events is that
Gitlab doesn't send Pipeline events if the merge request source is from
a fork. This means email notifications aren't sent for forks, which is
pretty useless.

Signed-off-by: Jeremy Cline <jcline@redhat.com>