Allow the subject prefix to be empty in forms and default to ""
The help text indicated the default prefix would be "PATCH" (a
roundabout way to say the empty string) but that doesn't match reality.
What's worse, the field was not marked blank so it's required in web
forms. Fix this and add a migration to set the default to the empty
string.

Signed-off-by: Jeremy Cline <jcline@redhat.com>